requirements_backend.txt
# FastAPI Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.13.1  # Database migrations

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
shap==0.44.1  # Explainable AI
joblib==1.3.2  # Model serialization

# External APIs
googlemaps==4.10.0
requests==2.31.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Security & Environment
python-dotenv==1.0.1
python-multipart==0.0.6  # File uploads
passlib[bcrypt]==1.7.4  # Password hashing (future auth)
python-jose[cryptography]==3.3.0  # JWT tokens (future auth)

# Utilities
python-dateutil==2.8.2
pytz==2023.3
